requirements.txt
elasticsearch==2.3.0
elasticsearch-dsl==2.0.0
Flask==0.10.1
Flask-Elasticsearch==0.2.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-dateutil==2.5.2
requests==2.9.1
six==1.10.0
urllib3==1.14
Werkzeug==0.11.5
Flask-Compress==1.3.0
flask_assets
progressbar2
